drivers: dstrx3: Fix implicit conversion from float to double
In C, a literal value like "2.5" is considered a double. When you
multiply a float by a double, implicit promotion occurs, performing
the calculation in 64 bits. This unnecessarily consumes CPU power on
MCUs.

   [118/166] Building C object .../sc_dstrx3.c.obj
   In file included from .../logging/log.h:11,
                    from .../sc_dstrx3.c:9:
   .../sc_dstrx3.c: In function 'sc_dstrx3_get_hk_telemetry':
   .../sc_dstrx3.c:132:83: warning: implicit conversion from 'float'
   to 'double' to match other operand of binary expression [-Wdouble-promotion]
     132 |         LOG_DBG("VOLTAGE          : %f [v] (raw: %d)", ((float)hk->voltage / 256) * 2.5,
         |                                                                                   ^

The goal is to perform all calculations under float. Then, for this
particular case, convert it to double because the value is passed to a
variadic function, z_log_minimal_printk().

Related resources:
- https://en.cppreference.com/w/c/language/conversion
- Link to PR: https://github.com/zephyrproject-rtos/zephyr/pull/57154

+++ zephyr/drivers/misc/sc_dstrx3.c
@@ -129,7 +129,7 @@ int sc_dstrx3_get_hk_telemetry(const struct device *dev, struct sc_dstrx3_hk *hk
 	LOG_DBG("RSSI             : %d", hk->rssi);
 	LOG_DBG("RCV_FREQ         : %d", hk->rcv_freq);
 	LOG_DBG("TEMPERATURE      : %d", hk->temperature);
-	LOG_DBG("VOLTAGE          : %f [v] (raw: %d)", (float)((hk->voltage / 256) * 2.5),
+	LOG_DBG("VOLTAGE          : %f [v] (raw: %d)", (double)(hk->voltage / 256.0f * 2.5f),
 		hk->voltage);
 	LOG_DBG("TX_PWR           : %d", hk->tx_power);
 	LOG_DBG("CARRIER_LOCK     : %d", hk->carrier_lock);
